🐛 Fix(#94): 데이터베이스 커넥션 반환 타이밍 수정 (홈, 최근 검색어, 쇼츠)

+++ src/research/research.service.js
@@ -1,7 +1,14 @@
-import { deleteRecentSearch, getRecentResearch } from "./research.dao.js";
+import { BaseError } from "../../config/error.js";
+import { status } from "../../config/response.status.js";
+import { deleteRecentSearch, getRecentResearch, getResearchUserId } from "./research.dao.js";
 
 // 최근 검색어 삭제
 export const deleteSearchService = async (research_id, user_id) => {
+    const researchUserId = await getResearchUserId(research_id);
+
+    if (user_id !== researchUserId) {
+        throw new BaseError(status.UNAUTHORIZED);
+    }
     return await deleteRecentSearch(research_id, user_id);
 }
